safari-600.7-branch/LayoutTests/sputnik/Conformance/13_Function_Definition/S13_A7_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Invalid escape in identifier: '\'
S13_A7_T3

PASS Expected parsing failure

TEST COMPLETE
